Convert bot to spring-boot app.
