+++ build.gradle
@@ -2,23 +2,30 @@ group 'ink.abb'
 version '0.3.0'
 
 buildscript {
-    ext.kotlin_version = '1.0.3'
+    ext {
+        kotlin_version = '1.0.3'
+        springBootVersion = '1.3.6.RELEASE'
+    }
 
     repositories {
         mavenCentral()
     }
     dependencies {
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}"
     }
 }
 
 apply plugin: 'java'
 apply plugin: 'kotlin'
 apply plugin: 'application'
+apply plugin: 'eclipse'
+apply plugin: 'spring-boot'
 
 mainClassName = 'ink.abb.pogo.scraper.MainKt'
 
-sourceCompatibility = 1.5
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 repositories {
     mavenCentral()
@@ -43,6 +50,9 @@ dependencies {
     compile group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.5.2'
 
     compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    compile('org.springframework.boot:spring-boot-starter')
+
+    testCompile('org.springframework.boot:spring-boot-starter-test')
     testCompile group: 'junit', name: 'junit', version: '4.11'
 }
 task fatJar(type: Jar) {
@@ -56,10 +66,20 @@ task fatJar(type: Jar) {
     with jar
 }
 jar {
+    baseName = 'PokemonGoBot'
+    version = '0.0.1-SNAPSHOT'
+
     manifest {
         attributes(
                 'Class-Path': configurations.compile.collect { it.getName() }.join(' '),
                 'Main-Class': mainClassName
         )
+    }
+}
+
+eclipse {
+    classpath {
+        containers.remove('org.eclipse.jdt.launching.JRE_CONTAINER')
+        containers 'org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8'
     }
 }
