+++ src/main/kotlin/ink/abb/pogo/scraper/Main.kt
@@ -8,100 +8,37 @@
 
 package ink.abb.pogo.scraper
 
-import POGOProtos.Networking.Envelopes.RequestEnvelopeOuterClass
 import com.pokegoapi.api.PokemonGo
-import com.pokegoapi.auth.GoogleLogin
-import com.pokegoapi.auth.PtcLogin
-import okhttp3.OkHttpClient
-import java.io.FileInputStream
-import java.net.InetSocketAddress
-import java.net.Proxy
-import java.security.cert.X509Certificate
-import java.util.*
-import java.util.concurrent.TimeUnit
-import javax.net.ssl.SSLContext
-import javax.net.ssl.TrustManager
-import javax.security.cert.CertificateException
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import javax.annotation.PostConstruct
 
 /**
  * Allow all certificate to debug with https://github.com/bettse/mitmdump_decoder
  */
-fun allowProxy(builder: OkHttpClient.Builder) {
-    builder.proxy(Proxy(Proxy.Type.HTTP, InetSocketAddress.createUnresolved("localhost", 8888)))
-    val trustAllCerts = arrayOf<TrustManager>(object : javax.net.ssl.X509TrustManager {
-        override fun getAcceptedIssuers(): Array<out X509Certificate> {
-            return emptyArray()
+@SpringBootApplication
+open class Main {
+
+    @Autowired
+    lateinit var settings: Settings
+
+    @Autowired
+    lateinit var api: PokemonGo
+
+    @PostConstruct
+    fun createBot() {
+        print("Getting profile data from pogo server")
+        while (api.playerProfile == null) {
+            print(".")
+            Thread.sleep(1000)
         }
+        println(".")
 
-        @Throws(CertificateException::class)
-        override fun checkClientTrusted(chain: Array<X509Certificate>, authType: String) {
-        }
-
-        @Throws(CertificateException::class)
-        override fun checkServerTrusted(chain: Array<X509Certificate>, authType: String) {
-        }
-    })
-
-    // Install the all-trusting trust manager
-    val sslContext = SSLContext.getInstance("SSL")
-    sslContext.init(null, trustAllCerts, java.security.SecureRandom())
-    // Create an ssl socket factory with our all-trusting manager
-    val sslSocketFactory = sslContext.socketFactory
-
-    builder.sslSocketFactory(sslSocketFactory)
-    builder.hostnameVerifier { hostname, session -> true }
+        Bot(api, settings).run()
+    }
 }
 
 fun main(args: Array<String>) {
-    val builder = OkHttpClient.Builder()
-    // allowProxy(builder)
-    builder.connectTimeout(60, TimeUnit.SECONDS)
-    builder.readTimeout(60, TimeUnit.SECONDS)
-    builder.writeTimeout(60, TimeUnit.SECONDS)
-    val http = builder.build()
-
-    val properties = Properties()
-    FileInputStream("config.properties").use {
-        properties.load(it)
-    }
-
-    val settings = Settings(properties)
-
-    val username = settings.username
-    val password = settings.password
-
-    val token = settings.token
-
-    println("Logging in to game server...")
-    val auth: RequestEnvelopeOuterClass.RequestEnvelope.AuthInfo
-
-    auth = if (token.isBlank()) {
-        if (username.contains('@')) {
-            GoogleLogin(http).login(username, password)
-        } else {
-            PtcLogin(http).login(username, password)
-        }
-    } else {
-        if (token.contains("pokemon.com")) {
-            PtcLogin(http).login(token)
-        } else {
-            GoogleLogin(http).login(token)
-        }
-    }
-
-    println("Logged in as $username with token ${auth.token.contents}")
-
-    if (token.isBlank()) {
-        println("Set this token in your config to log in directly")
-    }
-    val api = PokemonGo(auth, http)
-
-    print("Getting profile data from pogo server")
-    while (api.playerProfile == null) {
-        print(".")
-        Thread.sleep(1000)
-    }
-    println(".")
-
-    Bot(api, settings).run()
+    SpringApplication.run(Main::class.java, *args)
 }
